chore: wrongFormula 에러 던지는 조건을 Double로 형변환 가능한지로 변경

+++ Calculator/Calculator/Model/ExpressionParser.swift
@@ -4,7 +4,7 @@ enum ExpressionParser {
         let splitedInput = componentByOperators(from: input)
 
         for index in 0...splitedInput.count/2 {
-            guard Int(splitedInput[index*2]) != nil else {
+            guard Double(splitedInput[index*2]) != nil else {
                 throw CalculatorError.wrongFormula
             }
         }
